+++ packages/paypal-js/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 9.5.0
+
+### Minor Changes
+
+- a5d2818: Adding Apple Pay payment component types to v6 SDK
+
+## 9.4.1
+
+### Patch Changes
+
+- 10258ca: chore(types): rename v6 subscription onApprove type
+- 1b44c19: remove unused files and minor doc updates
+
 ## 9.4.0
 
 ### Minor Changes